requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1
Pillow==10.2.0
Flask-SQLAlchemy==3.1.1 